backend/auth_service/pyproject.toml
[project]
name = "auth-service"
version = "1.0.0"
description = "Authentication and user management service for AI-Native Textbook Platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "sqlalchemy>=2.0.36",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.28.0",
    "alembic>=1.14.0",
    "python-multipart>=0.0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.alembic]
script_path = "alembic"
file_template = "%%(year)d_%%(month).2d_%%(day).2d_%%(rev)s_%%(slug)s"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
